hf_manager/requirements.txt
# Core dependencies
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.5
langchain-deepseek>=0.0.1
langchain-groq>=0.1.0
langgraph>=0.0.25

# Data processing and analysis
polars>=0.20.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
colorama>=0.4.6
questionary>=2.0.0
tabulate>=0.9.0
pydantic>=2.0.0

# For LLM model integration
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0